+++ backend/django_project/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     # Local apps
     "users.apps.UsersConfig",
+    "bookmarks.apps.BookmarksConfig",
 ]
 
 MIDDLEWARE = [