+++ changelog.md
@@ -1,6 +1,7 @@
 ### 05.04.2023, Wed
 
 - 12:50 - Created `users` app with `CustomUser` model.
+- 14:00 - Created `bookmarks` app with `Folder` and `Bookmark` models.
 
 ### 03.04.2023, Mo